Issue #127: added SpriteSet class

+++ GeometryApp/src/com/js/geometry/Rect.java
@@ -132,25 +132,25 @@ public class Rect {
 		height = ey - y;
 	}
 
-	public double distanceFrom(Point pt) {
-		double dist = 0;
+	public float distanceFrom(Point pt) {
+		float dist = 0;
 		if (!contains(pt)) {
-			double dx = x - pt.x;
+			float dx = x - pt.x;
 			if (dx < 0)
 				dx = pt.x - endX();
 			if (dx < 0)
 				dx = 0;
-			double dy = y - pt.y;
+			float dy = y - pt.y;
 			if (dy < 0)
 				dy = pt.y - endY();
 			if (dy < 0)
 				dy = 0;
-			dist = Math.sqrt(dx * dx + dy * dy);
+			dist = (float) Math.sqrt(dx * dx + dy * dy);
 		}
 		return dist;
 	}
 
-	public void translate(double dx, double dy) {
+	public void translate(float dx, float dy) {
 		x += dx;
 		y += dy;
 	}
@@ -172,7 +172,7 @@ public class Rect {
 	 * 
 	 * @param f
 	 */
-	public void scale(double f) {
+	public void scale(float f) {
 		x *= f;
 		y *= f;
 		width *= f;
